# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadOut": {
        "description": "One attached file as the intake panel shows it (`REQ-DOC-003`).",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "state": {
            "$ref": "#/components/schemas/UploadState"
          }
        },
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "state",
          "error",
          "chunk_count",
          "created_at"
        ],
        "title": "UploadOut",
        "type": "object"
      },
      "UploadState": {
        "description": "Document processing lifecycle (`REQ-DOC-003..004`).",
        "enum": [
          "pending",
          "processing",
          "ready",
          "failed"
        ],
        "title": "UploadState",
        "type": "string"
      },
      "UploadTicket": {
        "description": "Where to PUT the bytes, and until when.\n\nThe URL is a short-lived write grant to one key (`DEC-12`). The API never\nsees the file: the browser uploads directly, which is implementation plan\n\u00a710 and the reason a 25 MB file does not have to fit through a request\nhandler.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "upload_id": {
            "format": "uuid",
            "title": "Upload Id",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "upload_id",
          "url",
          "expires_at",
          "content_type"
        ],
        "title": "UploadTicket",
        "type": "object"
      },
      "UploadTicketRequest": {
        "description": "Ask for permission to upload one file.\n\nThe size is what the *client* says it is. It is checked here so an\noversized file is refused before it is transferred rather than after\n(`REQ-DOC-010 AC-2`), and checked again at completion against the object\nthat actually arrived, because a declared size is not a limit.",
        "properties": {
          "content_type": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Content Type",
            "type": "string"
          },
          "filename": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename",
            "type": "string"
          },
          "size_bytes": {
            "exclusiveMinimum": 0.0,
            "title": "Size Bytes",
            "type": "integer"
          }
        },
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "UploadTicketRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Turns a question into an evidence-backed report with sources, analysis, charts and follow-up.",
    "title": "ScrapR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/research/{session_id}/uploads": {
      "get": {
        "description": "Every file attached to this session, with its processing state.\n\nThis is what `REQ-DOC-003 AC-1` is rendered from. Polled by the intake\npanel while anything is still processing, which is why it is cheap: one\nindexed query and a count per row.",
        "operationId": "list_uploads_v1_research__session_id__uploads_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UploadOut"
                  },
                  "title": "Response List Uploads V1 Research  Session Id  Uploads Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Uploads",
        "tags": [
          "uploads"
        ]
      },
      "post": {
        "description": "Reserve a slot and sign a URL for one file.\n\n**201, not 200**: this creates the `uploads` row. The row exists before the\nbytes do, which is what lets the count limit hold \u2014 a client that presigns\nten tickets and uploads none has used its ten.",
        "operationId": "create_upload_ticket_v1_research__session_id__uploads_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicket"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Upload Ticket",
        "tags": [
          "uploads"
        ]
      }
    },
    "/v1/research/{session_id}/uploads/{upload_id}": {
      "delete": {
        "description": "Remove a file and its extracted text (`REQ-SEC-008 AC-2`).\n\nThe chunks go with the row, immediately: they are the searchable copy, and\na deleted document that can still be retrieved as evidence is not deleted.\nThe object is removed next. If storage refuses, the row still goes \u2014 the\nuser's instruction is honoured and the orphaned object is a cleanup\nproblem, not a privacy one that waits on a retry.",
        "operationId": "delete_upload_v1_research__session_id__uploads__upload_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Upload",
        "tags": [
          "uploads"
        ]
      }
    },
    "/v1/research/{session_id}/uploads/{upload_id}/complete": {
      "post": {
        "description": "Confirm the bytes arrived, and queue the file for reading.\n\nThe size is read back from storage and re-checked, which is the second half\nof `REQ-DOC-010 AC-1`. A file that is over the limit \u2014 or that never\narrived \u2014 leaves no usable row behind.",
        "operationId": "complete_upload_v1_research__session_id__uploads__upload_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "upload_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Upload Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Upload",
        "tags": [
          "uploads"
        ]
      }
    },
...
  }
}